backend/requirements.txt
# Core FastAPI dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0

# Database dependencies
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.1

# Testing dependencies
pytest==7.4.3
pytest-cov==4.1.0
pytest-asyncio==0.21.1
hypothesis==6.92.1

# Development tools
black==23.11.0
isort==5.12.0
mypy==1.7.1

# Additional utilities
python-dotenv==1.0.0
httpx==0.25.2